requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
easyocr==1.7.1
email_validator==2.2.0
environs==9.5.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.112.2
fastapi-cli==0.0.5
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
imageio==2.35.1
importlib_metadata==8.4.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
JPype1==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwipiepy==0.18.0
kiwipiepy_model==0.18.0
kiwisolver==1.4.5
konlpy==0.6.0
langchain==0.2.15
langchain-community==0.2.14
langchain-core==0.2.36
langchain-huggingface==0.0.3
langchain-milvus==0.1.4
langchain-openai==0.1.23
langchain-teddynote==0.0.28
langchain-text-splitters==0.2.2
langsmith==0.1.107
lazy_loader==0.4
lxml==5.3.0
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
milvus-lite==2.4.9
mmh3==4.1.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
olefile==0.47
openai==1.43.0
opencv-python-headless==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
pinecone-text==0.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.4
protoc-gen-openapiv2==0.0.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyclipper==1.3.0.post5
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pymilvus==2.4.6
pyparsing==3.1.4
pytesseract==0.3.13
python-bidi==0.6.0
python-dateutil==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
rank-bm25==0.2.2
regex==2024.7.24
requests==2.32.3
rich==13.8.0
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.1
sentence-transformers==3.0.1
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
stack-data==0.6.2
starlette==0.38.2
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tifffile==2024.8.28
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
triton==3.0.0
typer==0.12.5
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1
wget==3.2
yarl==1.9.4
zipp==3.20.1
boto3==1.35.14
pyshorteners==1.0.1